+++ package.json
@@ -58,6 +58,7 @@
     },
     "description": "Hackolade plugin for Azure Cosmos DB with MongoDB API",
     "dependencies": {
+        "async": "3.2.6",
         "axios": "1.7.4",
         "bson": "1.1.5",
         "lodash": "4.17.21",
