+++ reverse_engineering/api.js
@@ -1,7 +1,7 @@
-'use strict';
-
-const CosmosClient = require('./CosmosClient');
+const _ = require('lodash');
 const bson = require('bson');
+const async = require('async');
+const CosmosClient = require('./CosmosClient');
 const connectionHelper = require('./helpers/connectionHelper');
 
 const ERROR_CONNECTION = 1;
@@ -71,7 +71,6 @@ module.exports = {
 	},
 
 	getDocumentKinds: function (connectionInfo, logger, cb, app) {
-		const async = app.require('async');
 		this.connect(connectionInfo, logger, (err, connection) => {
 			if (err) {
 				logger.log('error', err);
@@ -146,8 +145,6 @@ module.exports = {
 	},
 
 	getDbCollectionsNames: function (connectionInfo, logger, cb, app) {
-		const _ = app.require('lodash');
-		const async = app.require('async');
 		this.connect(connectionInfo, logger, (err, connection) => {
 			if (err) {
 				logger.log('error', err);
@@ -199,7 +196,6 @@ module.exports = {
 	},
 
 	getDbCollectionsData: function (data, logger, cb, app) {
-		const async = app.require('async');
 		let includeEmptyCollection = data.includeEmptyCollection;
 		let { recordSamplingSettings, fieldInference } = data;
 		logger.progress = logger.progress || (() => {});
@@ -655,7 +651,7 @@ function generateCustomInferSchema(documents, params) {
 					inferSchema.properties[prop]['samples'].indexOf(item[prop]) === -1 &&
 					inferSchema.properties[prop]['samples'].length < sampleSize
 				) {
-					inferSchema.properties[prop]['samples'].push(item[prop])
+					inferSchema.properties[prop]['samples'].push(item[prop]);
 				}
 
 				inferSchema.properties[prop]['type'] = typeOf(item[prop]);
